fix: heartbeat and id gen
